Добавляет обработчик для фильтров
